+++ src/helpers/promiseWait.ts
@@ -4,14 +4,40 @@
  * If Promise resolves or rejects in less than pointed time: waiting for rest time
  * @param promise Promise that we must wait for pointed time
  * @param ms min time that Promise must take
+ * @param [smartOrCallback=true] Allow not to wait if pointed promise is resolved immediately default-`true`
+ * @example
+ * let isPending = false
+ * promiseWait(Promise.resolve(), 300, ()=>isPending=true).finally(()=>isPending=false)
+ * // OR
+ * isPending = true;
+ * promiseWait(Promise.resolve(), 300, true).finally(()=>isPending=false)
  */
-export default function promiseWait<T>(promise: Promise<T>, ms: number): Promise<T> {
+export default function promiseWait<T>(
+  promise: Promise<T>,
+  ms: number,
+  smartOrCallback: boolean | Function = true
+): Promise<T> {
   let catchErr: Error;
-  return new Promise((resolve, reject) => {
-    setTimeout(() => (catchErr ? reject(catchErr) : resolve(promise)), ms);
-    promise.catch((err) => {
+  let isResolved = false;
+  promise
+    .catch((err) => {
       catchErr = err;
       return err;
+    })
+    .finally(() => {
+      isResolved = true;
     });
+
+  return new Promise((resolve, reject) => {
+    const end = () => (catchErr ? reject(catchErr) : resolve(promise));
+    const a = setTimeout(end, ms);
+    smartOrCallback &&
+      setTimeout(() => {
+        if (isResolved) {
+          clearTimeout(a);
+          end();
+          smartOrCallback instanceof Function && smartOrCallback();
+        }
+      });
   });
 }