+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   - Added TextControl (desc. coming soon...)
   - Added SelectControl (desc. coming soon...)
 - Helper [nestedProperty](README.md#helpers): added option `out.hasProp`
+- Helper [promiseWait](README.md#helpers): added option `smartOrCallback` to prevent useless pending on already resolved promises
 
 - [`PopupElement`](README.md#popupelement)
   - Added $refresh() - to force update/recalc position when nested content is changed